[`flake8-type-checking`] Fixes `quote_type_expression`
