+++ crates/ruff_linter/src/rules/flake8_type_checking/rules/runtime_import_in_type_checking_block.rs
@@ -270,6 +270,7 @@ fn quote_imports(checker: &Checker, node_id: NodeId, imports: &[ImportBinding]) 
                         Some(quote_annotation(
                             reference.expression_id()?,
                             checker.semantic(),
+                            checker.locator(),
                             checker.stylist(),
                         ))
                     } else {
